+++ src/main/java/mainProgram/services/JobService.java
@@ -23,18 +23,18 @@ public class JobService implements BaseSearchService<Job> {
         this.productRepository = productRepository;
     }
 
-    public Job getJobById(Long id) {
+    public Job getJobById(int id) {
         return jobRepository.findById(Math.toIntExact(id)).orElseThrow(() -> new RuntimeException("Job not found"));
     }
 
     /// Find the parts associated with a repair
-    public List<JobPart> getPartsForJob(Long jobId) {
+    public List<JobPart> getPartsForJob(int jobId) {
         return jobPartRepository.findByJobId(jobId);
     }
 
 
     /// Add a new product to a repair, using the JobPart join-table
-    public void addProductToRepair(Long repairId, Long productId, int quantity) {
+    public void addProductToRepair(int repairId, int productId, int quantity) {
         Job repair = getJobById(repairId);
         Product product = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("Product not found"));
